+++ earthmover/node.py
@@ -1,18 +1,20 @@
 import abc
 import dask
-import dask.dataframe as dd
 import jinja2
 import pandas as pd
 
 from dask.diagnostics import ProgressBar
-from typing import List
 
-from earthmover.yaml_parser import YamlMapping
 from earthmover import util
 
+from typing import Dict, List, Tuple, Optional, Union
 from typing import TYPE_CHECKING
 if TYPE_CHECKING:
+    from dask.dataframe.core import DataFrame
     from earthmover.earthmover import Earthmover
+    from earthmover.error_handler import ErrorHandler
+    from earthmover.yaml_parser import YamlMapping
+    from logging import Logger
 
 
 class Node:
@@ -20,17 +22,21 @@ class Node:
 
     """
     type: str = None
-    allowed_configs: tuple = ('debug', 'expect', 'show_progress', 'chunksize')
+    allowed_configs: Tuple[str] = ('debug', 'expect', 'show_progress', 'repartition',)
 
-    def __init__(self, name: str, config: YamlMapping, *, earthmover: 'Earthmover'):
-        self.name = name
-        self.config = config
+    def __init__(self, name: str, config: 'YamlMapping', *, earthmover: 'Earthmover'):
+        self.name: str = name
+        self.config: 'YamlMapping' = config
 
-        self.earthmover = earthmover
-        self.logger = earthmover.logger
-        self.error_handler = earthmover.error_handler
+        self.earthmover: 'Earthmover' = earthmover
+        self.logger: 'Logger' = earthmover.logger
+        self.error_handler: 'ErrorHandler' = earthmover.error_handler
 
-        self.upstream_sources: dict = {}
+        self.error_handler.ctx.update(
+            file=self.earthmover.config_file, line=self.config.__line__, node=self, operation=None
+        )
+
+        self.upstream_sources: Dict[str, Optional['Node']] = {}
 
         self.data: 'DataFrame' = None
 
@@ -38,15 +44,15 @@ class Node:
         self.num_rows: int = None
         self.num_cols: int = None
 
-        self.expectations: list = None
+        self.expectations: List[str] = None
         self.debug: bool = False
 
-        # Customize internal Dask configs
-        self.chunksize = self.config.get('chunksize', self.earthmover.state_configs["chunksize"])
+        # Internal Dask configs
+        self.partition_size: Union[str, int] = self.config.get('repartition')
 
         # Optional variables for displaying progress and diagnostics.
-        self.show_progress = self.config.get('show_progress', self.earthmover.state_configs["show_progress"])
-        self.progress_bar = None
+        self.show_progress: bool = self.config.get('show_progress', self.earthmover.state_configs["show_progress"])
+        self.progress_bar: ProgressBar = ProgressBar(minimum=10, dt=5.0)  # Always instantiate, but only use if `show_progress is True`.
 
     @abc.abstractmethod
     def compile(self):
@@ -73,10 +79,10 @@ class Node:
         pass
 
     @abc.abstractmethod
-    def execute(self, **kwargs) -> 'DataFrame':
+    def execute(self, **kwargs):
         """
-        Node.execute()      :: Saves data into memory
-        Operation.execute() :: Does NOT save data into memory
+        Node.execute()          :: Saves data into memory
+        Operation.execute(data) :: Does NOT save data into memory
 
         :return:
         """
@@ -84,29 +90,33 @@ class Node:
             file=self.earthmover.config_file, line=self.config.__line__, node=self, operation=None
         )
 
+        # Turn on the progress bar manually.
         if self.show_progress:
             self.logger.info(f"Displaying progress for {self.type} node: {self.name}")
-            self.progress_bar = ProgressBar(dt=1.0)
             self.progress_bar.__enter__()  # Open context manager manually to avoid with-clause
 
         pass
 
-    def post_execute(self):
+    @abc.abstractmethod
+    def post_execute(self, **kwargs):
         """
         Function to run generic logic following execute.
 
-        1. Check the dataframe aligns with expectations
-        2. Prepare row and column counts for graphing
+        1. Complete any post-transformations to self.data (currently unused).
+        2. Check the dataframe aligns with expectations.
+        3. Prepare row and column counts for graphing.
+        4. Display row and column counts if debug is True.
 
         :return:
         """
+        # Close context manager manually to avoid with-clause.
         if self.show_progress:
-            self.progress_bar.__exit__(None, None, None)  # Close context manager manually to avoid with-clause
+            self.progress_bar.__exit__(None, None, None)
 
         self.check_expectations(self.expectations)
 
         # Get lazy row and column counts to display in graph.png.
-        if isinstance(self.data, (pd.Series, dd.Series)):
+        if isinstance(self.data, (pd.Series, dask.dataframe.Series)):
             self.num_rows, self.num_cols = self.data.size, 1
         else:
             self.num_rows, self.num_cols = self.data.shape
@@ -115,8 +125,10 @@ class Node:
             self.num_rows = dask.compute(self.num_rows)[0]
             self.logger.debug(
                 f"Node {self.name}: {self.num_rows} rows; {self.num_cols} columns\n"
-                f"Header: {self.data.columns}"
+                f"Header: {self.data.columns if hasattr(self.data, 'columns') else 'No header'}"
             )
+
+        pass
 
     def check_expectations(self, expectations: List[str]):
         """
@@ -148,3 +160,8 @@ class Node:
                     self.logger.info(
                         f"Assertion passed! {self.name}: {expectation}"
                     )
+
+    def opt_repartition(self, data: 'DataFrame'):
+        if self.partition_size:
+            data = data.repartition(partition_size=self.partition_size)
+        return data