+++ earthmover/nodes/destination.py
@@ -1,11 +1,14 @@
-import os
-
+import csv
 import jinja2
+import os
 import pandas as pd
 import re
 
 from earthmover.node import Node
 from earthmover import util
+
+from typing import Tuple
+
 
 class Destination(Node):
     """
@@ -13,14 +16,14 @@ class Destination(Node):
     """
     type: str = 'destination'
     mode: str = None  # Documents which class was chosen.
-    allowed_configs: tuple = ('debug', 'expect', 'show_progress', 'chunksize', 'source',)
+    allowed_configs: Tuple[str] = ('debug', 'expect', 'show_progress', 'repartition', 'source',)
 
     def __new__(cls, *args, **kwargs):
         return object.__new__(FileDestination)
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.source = self.error_handler.assert_get_key(self.config, 'source', dtype=str)
+        self.source: str = self.error_handler.assert_get_key(self.config, 'source', dtype=str)
         self.upstream_sources[self.source] = None
 
 
@@ -29,8 +32,8 @@ class FileDestination(Destination):
 
     """
     mode: str = 'file'
-    allowed_configs: tuple = (
-        'debug', 'expect', 'show_progress', 'chunksize', 'source',
+    allowed_configs: Tuple[str] = (
+        'debug', 'expect', 'show_progress', 'repartition', 'source',
         'template', 'extension', 'linearize', 'header', 'footer',
     )
 
@@ -94,11 +97,14 @@ class FileDestination(Destination):
             )
             raise
 
-    def execute(self):
+    def execute(self, **kwargs):
         """
+        There is a bug in Dask where where `dd.to_csv(mode='a', single_file=True)` fails.
+        This is resolved in 2023.8.1: https://docs.dask.org/en/stable/changelog.html#id7 
+
         :return:
         """
-        super().execute()
+        super().execute(**kwargs)
 
         # this renders each row without having to itertuples() (which is much slower)
         # (meta=... below is how we prevent dask warnings that it can't infer the output data type)
@@ -108,23 +114,27 @@ class FileDestination(Destination):
                 .map_partitions(lambda x: x.apply(self.render_row, axis=1), meta=pd.Series('str'))
         )
 
+        # Repartition before writing, if specified.
+        self.data = self.opt_repartition(self.data)
+
+        # Verify the output directory exists.
         os.makedirs(os.path.dirname(self.file), exist_ok=True)
-        with open(self.file, 'w', encoding='utf-8') as fp:
-            self.logger.debug(f"writing output file `{self.file}`...")
 
-            if self.header:
-                fp.write(self.header + "\n")
+        # Write the optional header, the JSON lines as CSV (for performance), and the optional footer.
+        self.data.to_csv(
+            filename=self.file, single_file=True, mode='wt', index=False,
+            header=[self.header] if self.header else False,  # We must write the header directly due to aforementioned bug.
+            escapechar="\x01", sep="\x02", quoting=csv.QUOTE_NONE,  # Pretend to be CSV to improve performance
+        )
 
-            for row in self.data.compute():
-                fp.write(row + "\n")
-
-            if self.footer:
+        if self.footer:
+            with open(self.file, 'a', encoding='utf-8') as fp:
                 fp.write(self.footer)
 
         self.logger.debug(f"output `{self.file}` written")
         self.size = os.path.getsize(self.file)
 
-    def render_row(self, row):
+    def render_row(self, row: pd.Series):
         _data_tuple = row.to_dict()
         _data_tuple["__row_data__"] = row