+++ earthmover/operations/column.py
@@ -1,27 +1,30 @@
 import csv
-import dask.dataframe as dd
-import jinja2
+import dask
+import pandas as pd
 import re
 import string
-import os
-import pandas as pd
 
 from earthmover.nodes.operation import Operation
 from earthmover import util
+
+from typing import Dict, List, Tuple
+from typing import TYPE_CHECKING
+if TYPE_CHECKING:
+    from dask.dataframe.core import DataFrame
 
 
 class AddColumnsOperation(Operation):
     """
 
     """
-    allowed_configs: tuple = (
-        'operation',
+    allowed_configs: Tuple[str] = (
+        'operation', 'repartition', 
         'columns',
     )
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.columns_dict: dict = None
+        self.columns_dict: Dict[str, str] = None
 
     def compile(self):
         """
@@ -31,7 +34,7 @@ class AddColumnsOperation(Operation):
         super().compile()
         self.columns_dict = self.error_handler.assert_get_key(self.config, 'columns', dtype=dict)
 
-    def execute(self, data: 'DataFrame', **kwargs):
+    def execute(self, data: 'DataFrame', **kwargs) -> 'DataFrame':
         """
 
         :return:
@@ -70,14 +73,14 @@ class ModifyColumnsOperation(Operation):
     """
 
     """
-    allowed_configs: tuple = (
-        'operation',
+    allowed_configs: Tuple[str] = (
+        'operation', 'repartition', 
         'columns',
     )
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.columns_dict: dict = None
+        self.columns_dict: Dict[str, str] = None
 
     def compile(self):
         """
@@ -87,7 +90,7 @@ class ModifyColumnsOperation(Operation):
         super().compile()
         self.columns_dict = self.error_handler.assert_get_key(self.config, 'columns', dtype=dict)
 
-    def execute(self, data: 'DataFrame', **kwargs):
+    def execute(self, data: 'DataFrame', **kwargs) -> 'DataFrame':
         """
 
         :return:
@@ -131,14 +134,14 @@ class DuplicateColumnsOperation(Operation):
     """
 
     """
-    allowed_configs: tuple = (
-        'operation',
+    allowed_configs: Tuple[str] = (
+        'operation', 'repartition', 
         'columns',
     )
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.columns_dict: dict = None
+        self.columns_dict: Dict[str, str] = None
 
     def compile(self):
         """
@@ -148,7 +151,7 @@ class DuplicateColumnsOperation(Operation):
         super().compile()
         self.columns_dict = self.error_handler.assert_get_key(self.config, 'columns', dtype=dict)
 
-    def execute(self, data: 'DataFrame', **kwargs):
+    def execute(self, data: 'DataFrame', **kwargs) -> 'DataFrame':
         """
 
         :return:
@@ -176,14 +179,14 @@ class RenameColumnsOperation(Operation):
     """
 
     """
-    allowed_configs: tuple = (
-        'operation',
+    allowed_configs: Tuple[str] = (
+        'operation', 'repartition', 
         'columns',
     )
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.columns_dict: dict = None
+        self.columns_dict: Dict[str, str] = None
 
     def compile(self):
         """
@@ -193,7 +196,7 @@ class RenameColumnsOperation(Operation):
         super().compile()
         self.columns_dict = self.error_handler.assert_get_key(self.config, 'columns', dtype=dict)
 
-    def execute(self, data: 'DataFrame', **kwargs):
+    def execute(self, data: 'DataFrame', **kwargs) -> 'DataFrame':
         """
 
         :return:
@@ -219,14 +222,14 @@ class DropColumnsOperation(Operation):
     """
 
     """
-    allowed_configs: tuple = (
-        'operation',
+    allowed_configs: Tuple[str] = (
+        'operation', 'repartition', 
         'columns',
     )
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.columns_to_drop: list = None
+        self.columns_to_drop: List[str] = None
 
     def compile(self):
         """
@@ -236,7 +239,7 @@ class DropColumnsOperation(Operation):
         super().compile()
         self.columns_to_drop = self.error_handler.assert_get_key(self.config, 'columns', dtype=list)
 
-    def execute(self, data: 'DataFrame', **kwargs):
+    def execute(self, data: 'DataFrame', **kwargs) -> 'DataFrame':
         """
 
         :return:
@@ -258,14 +261,14 @@ class KeepColumnsOperation(Operation):
     """
 
     """
-    allowed_configs: tuple = (
-        'operation',
+    allowed_configs: Tuple[str] = (
+        'operation', 'repartition', 
         'columns',
     )
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.header: list = None
+        self.header: List[str] = None
 
     def compile(self):
         """
@@ -276,7 +279,7 @@ class KeepColumnsOperation(Operation):
 
         self.header = self.error_handler.assert_get_key(self.config, 'columns', dtype=list)
 
-    def execute(self, data: 'DataFrame', **kwargs):
+    def execute(self, data: 'DataFrame', **kwargs) -> 'DataFrame':
         """
 
         :return:
@@ -298,14 +301,14 @@ class CombineColumnsOperation(Operation):
     """
 
     """
-    allowed_configs: tuple = (
-        'operation',
+    allowed_configs: Tuple[str] = (
+        'operation', 'repartition', 
         'columns', 'new_column', 'separator',
     )
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.columns_list: list = None
+        self.columns_list: List[str] = None
         self.new_column: str = None
         self.separator: str = None
 
@@ -321,7 +324,7 @@ class CombineColumnsOperation(Operation):
 
         self.separator = self.config.get('separator', "")
 
-    def execute(self, data: 'DataFrame', **kwargs):
+    def execute(self, data: 'DataFrame', **kwargs) -> 'DataFrame':
         """
 
         :return:
@@ -348,16 +351,16 @@ class MapValuesOperation(Operation):
     """
 
     """
-    allowed_configs: tuple = (
-        'operation',
+    allowed_configs: Tuple[str] = (
+        'operation', 'repartition', 
         'column', 'columns', 'mapping', 'map_file',
     )
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.columns_list: list = None
+        self.columns_list: List[str] = None
         self.map_file: str = None
-        self.mapping: dict = None
+        self.mapping: Dict[str, str] = None
 
     def compile(self):
         """
@@ -392,7 +395,7 @@ class MapValuesOperation(Operation):
             )
             raise
 
-    def execute(self, data: 'DataFrame', **kwargs):
+    def execute(self, data: 'DataFrame', **kwargs) -> 'DataFrame':
         """
 
         :return:
@@ -441,14 +444,14 @@ class DateFormatOperation(Operation):
     """
 
     """
-    allowed_configs: tuple = (
-        'operation',
+    allowed_configs: Tuple[str] = (
+        'operation', 'repartition', 
         'column', 'columns', 'from_format', 'to_format',
     )
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.columns_list: list = None
+        self.columns_list: List[str] = None
         self.from_format: str = None
         self.to_format: str = None
 
@@ -474,7 +477,7 @@ class DateFormatOperation(Operation):
 
         self.columns_list = _columns or [_column]  # `[None]` evaluates to True
 
-    def execute(self, data: 'DataFrame', **kwargs):
+    def execute(self, data: 'DataFrame', **kwargs) -> 'DataFrame':
         """
 
         :return:
@@ -490,7 +493,7 @@ class DateFormatOperation(Operation):
         for _column in self.columns_list:
             try:
                 data[_column] = (
-                    dd.to_datetime(data[_column], format=self.from_format)
+                    dask.dataframe.to_datetime(data[_column], format=self.from_format)
                         .dt.strftime(self.to_format)
                 )
 
@@ -507,8 +510,8 @@ class SnakeCaseColumnsOperation(Operation):
     """
 
     """
-    allowed_configs: tuple = (
-        'operation',
+    allowed_configs: Tuple[str] = (
+        'operation', 'repartition', 
     )
 
     def execute(self, data: 'DataFrame', **kwargs) -> 'DataFrame':