+++ earthmover/operations/groupby.py
@@ -3,24 +3,26 @@ import re
 
 from earthmover.nodes.operation import Operation
 
+from typing import Dict, List, Tuple
+from typing import TYPE_CHECKING
+if TYPE_CHECKING:
+    from dask.dataframe.core import DataFrame
 
 class GroupByWithCountOperation(Operation):
     """
 
     """
-    allowed_configs: tuple = (
-        'operation',
+    allowed_configs: Tuple[str] = (
+        'operation', 'repartition',  
         'group_by_columns', 'count_column',
     )
-
-
 
     GROUPED_COL_NAME = "____grouped_col____"
     GROUPED_COL_SEP = "_____"
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.group_by_columns: list = None
+        self.group_by_columns: List[str] = None
         self.count_column: str = None
 
     def compile(self):
@@ -32,7 +34,7 @@ class GroupByWithCountOperation(Operation):
         self.group_by_columns = self.error_handler.assert_get_key(self.config, 'group_by_columns', dtype=list)
         self.count_column     = self.error_handler.assert_get_key(self.config, 'count_column', dtype=str)
 
-    def execute(self, data: 'DataFrame', **kwargs):
+    def execute(self, data: 'DataFrame', **kwargs) -> 'DataFrame':
         """
 
         :return:
@@ -68,8 +70,8 @@ class GroupByWithAggOperation(Operation):
     """
 
     """
-    allowed_configs: tuple = (
-        'operation',
+    allowed_configs: Tuple[str] = (
+        'operation', 'repartition', 
         'group_by_columns', 'agg_column', 'separator',
     )
 
@@ -79,7 +81,7 @@ class GroupByWithAggOperation(Operation):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.group_by_columns: list = None
+        self.group_by_columns: List[str] = None
         self.agg_column: str = None
         self.separator: str = None
 
@@ -97,7 +99,7 @@ class GroupByWithAggOperation(Operation):
             required=False, default=self.DEFAULT_AGG_SEP
         )
 
-    def execute(self, data: 'DataFrame', **kwargs):
+    def execute(self, data: 'DataFrame', **kwargs) -> 'DataFrame':
         """
 
         :return:
@@ -131,8 +133,8 @@ class GroupByOperation(Operation):
     """
 
     """
-    allowed_configs: tuple = (
-        'operation',
+    allowed_configs: Tuple[str] = (
+        'operation', 'repartition', 
         'group_by_columns', 'create_columns',
     )
 
@@ -150,8 +152,8 @@ class GroupByOperation(Operation):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-        self.group_by_columns: list = None
-        self.create_columns_dict: dict = None
+        self.group_by_columns: List[str] = None
+        self.create_columns_dict: Dict[str, str] = None
 
     def compile(self):
         """
@@ -162,7 +164,7 @@ class GroupByOperation(Operation):
         self.group_by_columns    = self.error_handler.assert_get_key(self.config, 'group_by_columns', dtype=list)
         self.create_columns_dict = self.error_handler.assert_get_key(self.config, 'create_columns', dtype=dict)
 
-    def execute(self, data: 'DataFrame', **kwargs):
+    def execute(self, data: 'DataFrame', **kwargs) -> 'DataFrame':
         """
         Note: There is a bug in Dask Groupby operations.
         Index columns are overwritten by 'index' after index reset.