+++ earthmover/util.py
@@ -1,12 +1,14 @@
 import jinja2
 import hashlib
-import logging
 import os
 
-from typing import Any, Optional
 from sys import exc_info
 
-from earthmover.error_handler import ErrorHandler
+from typing import Optional
+from typing import TYPE_CHECKING
+if TYPE_CHECKING:
+    from earthmover.error_handler import ErrorHandler
+    from pandas import Series
 
 
 def human_time(seconds: int) -> str:
@@ -29,8 +31,8 @@ def human_time(seconds: int) -> str:
     return str(round(seconds/86400)) + " days"
 
 
-def human_size(bytes, units=['B','KB','MB','GB','TB', 'PB', 'EB']):
-    return str(bytes) + units[0] if bytes < 1024 else human_size(bytes>>10, units[1:])
+def human_size(bytes_: int, units=('B','KB','MB','GB','TB', 'PB', 'EB')):
+    return str(bytes_) + units[0] if bytes_ < 1024 else human_size(bytes_>>10, units[1:])
 
 def get_sep(file: str) -> Optional[str]:
     """
@@ -67,7 +69,7 @@ def contains_jinja(string: str) -> bool:
         return False
 
 
-def render_jinja_template(row, template: jinja2.Template, template_str: str, *, error_handler: ErrorHandler) -> str:
+def render_jinja_template(row: 'Series', template: jinja2.Template, template_str: str, *, error_handler: 'ErrorHandler') -> str:
     """
 
     :param row:
@@ -99,10 +101,10 @@ def jinja2_template_error_lineno():
     function based on https://stackoverflow.com/questions/26967433/how-to-get-line-number-causing-an-exception-other-than-templatesyntaxerror-in
     :return: int lineno
     """
-    type, value, tb = exc_info()
+    type_, value, tb = exc_info()
 
     # skip non-Jinja errors
-    if not issubclass(type, jinja2.TemplateError):
+    if not issubclass(type_, jinja2.TemplateError):
         return None
 
     # one particular Exception type has a lineno built in - grab it!
@@ -124,29 +126,8 @@ def build_jinja_template(template_string: str, macros: str = ""):
     """
     template = jinja2.Environment(
         loader=jinja2.FileSystemLoader(os.path.dirname('./'))
-    ).from_string(macros + template_string)
+    ).from_string(macros.strip() + template_string)
 
     template.globals['md5'] = lambda x: hashlib.md5(x.encode('utf-8')).hexdigest()
 
     return template
-
-
-# TODO: This has to be a NULL variable, without being None!
-__UNDEFINED = "[[UNDEFINED]]"
-
-def assert_get_key(obj: dict, key: str, *, default: Optional[Any] = __UNDEFINED, dtype: Any = object):
-    value = obj.get(key, default)
-
-    if value == __UNDEFINED:
-        logging.critical(
-            f"YAML parse error: Field not defined: {key}."
-        )
-
-    if not isinstance(value, dtype):
-        logging.critical(
-            f"YAML parse error: Field does not match expected datatype: {key}\n"
-            f"    Expected: {dtype}\n"
-            f"    Received: {value}"
-        )
-
-    return value