+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+### v0.2.0
+<details>
+<summary>Released 2023-09-11</summary>
+* breaking change: remove `source` as Operation config and move to Transformation; this simplifies templates and reduces memory usage
+* breaking change: `version: 2` required in Earthmover YAML files 
+* feature: `SnakeCaseColumnsOperation` converts all columns to snake_case
+* feature: `show_progress` can be turned on globally in `config` or locally in any Source, Transformation, or Destination to display a progress bar
+* feature: `repartition` can be turned on in any applicable `Node` to alter Dask partition-sizes post-execute
+* feature: improve performance when writing Destination files
+* feature: improved Earthmover YAML-parsing and config-retrieval
+* internal: rename `YamlEnvironmentJinjaLoader` to `JinjaEnvironmentYamlLoader` for better transparency of use
+* internal: simplify Earthmover.build_graph()
+* internal: unify Jinja rendering into a single util function, instead of redeclaring across project
+* internal: unify `Node.verify()` into `Node.execute()` for improved code legibility
+* internal: improve attribute declarations across project
+* internal: improve type-hinting and doc-strings across project
+* bugfix: refactor SqlSource to be compatible with SQLAlchemy 2.x
+
+</details>
+
 ### v0.1.6
 <details>
 <summary>Released 2023-07-11</summary>